+++ packages/react/src/components/UpsellingKit/exports.tsx
@@ -1,2 +1,4 @@
 export * from "./ProductCard"
 export * from "./ProductWidget"
+export * from "./UpsellingButton"
+export * from "./UpsellRequestResponseDialog"